pyproject.toml
[project]
name = "pyqt"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.7"
dependencies = [
    "pyqt5>=5.15.10",
    "pyyaml>=6.0.1",
    "scikit-learn>=1.0.2",
    "ultralytics>=8.0.145",
]
[tool.uv]
constraint-dependencies = ["pyqt5-qt5 <=5.15.2"]
